SimpleWeather: add a fallback for getting weather data
* Query the DB to get weather by using lat,long from weather data (locationblob)
* Add null check to avoid exceptions
* Cleanup updateLocationQuery method (avoid calling API/change query directly)
* Format lat,long string using .toString method instead of String.format directly

+++ shared_resources/src/main/java/com/thewizrd/shared_resources/weatherdata/Location.java
@@ -68,8 +68,8 @@ public class Location {
     public Location(com.thewizrd.shared_resources.weatherdata.openweather.ForecastRootobject root) {
         // Use location name from location provider
         name = null;
-        latitude = Float.toString(root.getCity().getCoord().getLat());
-        longitude = Float.toString(root.getCity().getCoord().getLon());
+        latitude = Double.toString(root.getCity().getCoord().getLat());
+        longitude = Double.toString(root.getCity().getCoord().getLon());
         tzOffset = ZoneOffset.UTC;
         tzShort = "UTC";
     }
@@ -86,8 +86,8 @@ public class Location {
     public Location(com.thewizrd.shared_resources.weatherdata.here.LocationItem location) {
         // Use location name from location provider
         name = null;
-        latitude = Float.toString(location.getLatitude());
-        longitude = Float.toString(location.getLongitude());
+        latitude = Double.toString(location.getLatitude());
+        longitude = Double.toString(location.getLongitude());
         tzOffset = ZoneOffset.UTC;
         tzShort = "UTC";
     }
